moved previous version of common into old_common.py (based on Peskin and initial studies)
